refactor: cache cmdutil.Factory to improve syncronization performance
